booth-3-record/backend/pyproject.toml
[project]
name = "spinat-booth-3-record"
version = "0.1.0"
description = "spinat 릴스 Booth 3 — Recording clone of booth-1 for pre-event triage (see archive/pump-game for the original booth-3 game code)"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi==0.110.1",
    "uvicorn[standard]==0.29.0",
    "python-multipart==0.0.9",
    "boto3==1.34.84",
    "qrcode[pil]==7.4.2",
    "Pillow==10.3.0",
    "PyYAML==6.0.1",
    "python-dotenv==1.0.1",
    "aiofiles==23.2.1",
]

[dependency-groups]
dev = [
    "pytest==8.1.1",
    "httpx==0.27.0",
]

[tool.uv]
package = false
